+++ cx/config.go
@@ -4,6 +4,13 @@ import (
 	"os"
 )
 
+// Initializing `CXProgram` structure where packages, structs, functions and
+// global variables that belong to core packages are stored.
+func init() {
+	prgrm := CXProgram{Packages: make([]*CXPackage, 0)}
+	PROGRAM = &prgrm
+}
+
 var InREPL bool = false
 var FoundCompileErrors bool
 
@@ -105,6 +112,7 @@ const (
 	DECL_STRUCT          // 4
 	DECL_INDEXING        // 5
 	DECL_BASIC           // 6
+	DECL_FUNC            // 7
 )
 
 // create a new scope or return to the previous scope
@@ -142,6 +150,7 @@ const (
 	TYPE_UI16
 	TYPE_UI32
 	TYPE_UI64
+	TYPE_FUNC
 
 	TYPE_CUSTOM
 	TYPE_POINTER
@@ -167,6 +176,7 @@ var TypeCodes map[string]int = map[string]int{
 	"ui32":  TYPE_UI32,
 	"ui64":  TYPE_UI64,
 	"und":   TYPE_UNDEFINED,
+	"func":  TYPE_FUNC,
 }
 
 var TypeNames map[int]string = map[int]string{
@@ -184,6 +194,7 @@ var TypeNames map[int]string = map[int]string{
 	TYPE_UI16:       "ui16",
 	TYPE_UI32:       "ui32",
 	TYPE_UI64:       "ui64",
+	TYPE_FUNC:       "func",
 	TYPE_UNDEFINED:  "und",
 }