+++ cxgo/actions/literals.go
@@ -1,13 +1,13 @@
 package actions
 
 import (
-	"github.com/amherag/skycoin/src/cipher/encoder"
+	"github.com/SkycoinProject/skycoin/src/cipher/encoder"
 
 	. "github.com/SkycoinProject/cx/cx"
 )
 
 // SliceLiteralExpression handles literal expressions by converting it to a series of `append` expressions.
-func SliceLiteralExpression(sliceDim []int, typSpec int, exprs []*CXExpression) []*CXExpression {
+func SliceLiteralExpression(typSpec int, exprs []*CXExpression) []*CXExpression {
 	var result []*CXExpression
 
 	pkg, err := PRGRM.GetCurrentPackage()
@@ -21,10 +21,8 @@ func SliceLiteralExpression(sliceDim []int, typSpec int, exprs []*CXExpression) 
 	slcVarExpr := MakeExpression(nil, CurrentFile, LineNo)
 	slcVarExpr.Package = pkg
 	slcVar := MakeArgument(symName, CurrentFile, LineNo)
-	slcVar = DeclarationSpecifiers(slcVar, sliceDim, DECL_SLICE)
 	slcVar.AddType(TypeNames[typSpec])
-
-	// slcVar.IsSlice = true
+	slcVar = DeclarationSpecifiers(slcVar, []int{0}, DECL_SLICE)
 
 	slcVar.TotalSize = TYPE_POINTER_SIZE
 
@@ -39,10 +37,8 @@ func SliceLiteralExpression(sliceDim []int, typSpec int, exprs []*CXExpression) 
 		if expr.IsArrayLiteral {
 			symInp := MakeArgument(symName, CurrentFile, LineNo).AddType(TypeNames[typSpec])
 			symInp.Package = pkg
-			symInp.Lengths = sliceDim
 			symOut := MakeArgument(symName, CurrentFile, LineNo).AddType(TypeNames[typSpec])
 			symOut.Package = pkg
-			symOut.Lengths = sliceDim
 
 			endPointsCounter++
 
@@ -61,39 +57,24 @@ func SliceLiteralExpression(sliceDim []int, typSpec int, exprs []*CXExpression) 
 				// We need to create a temporary variable to hold the result of the
 				// nested expressions. Then use that variable as part of the slice literal.
 				out := MakeArgument(MakeGenSym(LOCAL_PREFIX), expr.FileName, expr.FileLine)
-				out = DeclarationSpecifiers(out, sliceDim[:len(sliceDim)-1], DECL_SLICE)
-				expr.IsArrayLiteral = false
 				outArg := getOutputType(expr)
 				out.AddType(TypeNames[outArg.Type])
 				out.CustomType = outArg.CustomType
 				out.Size = outArg.Size
 				out.TotalSize = GetSize(outArg)
-				out.Lengths = sliceDim
 				out.PreviouslyDeclared = true
-
-				// Declaring `out`
-				declExpr := MakeExpression(nil, CurrentFile, LineNo)
-				declExpr.Package = pkg
-				declExpr.AddOutput(out)
-				result = append(result, declExpr)
 
 				expr.Outputs = nil
 				expr.AddOutput(out)
 				result = append(result, expr)
 
-				// symExpr.Operator = expr.Operator
 				symExpr.Operator = Natives[OP_APPEND]
 
 				symExpr.Inputs = nil
 				symExpr.Inputs = append(symExpr.Inputs, symInp)
-				// symExpr.Inputs = append(symExpr.Inputs, expr.Inputs...)
 				symExpr.Inputs = append(symExpr.Inputs, out)
-
-				// hack to get the correct lengths below
-				// expr.Outputs = append(expr.Outputs, symInp)
 			}
 
-			// result = append(result, expr)
 			result = append(result, symExpr)
 
 			symInp.TotalSize = TYPE_POINTER_SIZE
@@ -101,37 +82,27 @@ func SliceLiteralExpression(sliceDim []int, typSpec int, exprs []*CXExpression) 
 		} else {
 			result = append(result, expr)
 		}
+		expr.IsArrayLiteral = false
 	}
 
 	symNameOutput := MakeGenSym(LOCAL_PREFIX)
 
 	symOutput := MakeArgument(symNameOutput, CurrentFile, LineNo).AddType(TypeNames[typSpec])
-	// symOutput.PassBy = PASSBY_REFERENCE
 	symOutput.IsSlice = true
 	symOutput.Package = pkg
 	symOutput.PreviouslyDeclared = true
 
-	// symOutput.DeclarationSpecifiers = append(symOutput.DeclarationSpecifiers, DECL_ARRAY)
-
 	symInput := MakeArgument(symName, CurrentFile, LineNo).AddType(TypeNames[typSpec])
-	// symInput.DereferenceOperations = append(symInput.DereferenceOperations, DEREF_POINTER)
 	symInput.IsSlice = true
 	symInput.Package = pkg
-	// symInput.PassBy = PASSBY_REFERENCE
 
 	symInput.TotalSize = TYPE_POINTER_SIZE
-	symInput.Lengths = sliceDim
 	symOutput.TotalSize = TYPE_POINTER_SIZE
-	symOutput.Lengths = sliceDim
 
 	symExpr := MakeExpression(Natives[OP_IDENTITY], CurrentFile, LineNo)
 	symExpr.Package = pkg
 	symExpr.Outputs = append(symExpr.Outputs, symOutput)
 	symExpr.Inputs = append(symExpr.Inputs, symInput)
-
-	// symExpr.IsArrayLiteral = true
-
-	// symOutput.SynonymousTo = symInput.Name
 
 	// marking the output so multidimensional arrays identify the expressions
 	result = append(result, symExpr)